+++ src/Phormium/Filter/CompositeFilter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phormium;
+namespace Phormium\Filter;
 
 /**
  * A filter which consists of several Filter objects which are joined by an